+++ lib/l10n/translations/tlh.dart
@@ -1756,6 +1756,9 @@ const _dutchSourceTlh = {
 };
 
 const _dutchSourceAddTlh = <String, String>{
+  'Norm en prestatie': 'ngoq je qav',
+  'Norm per rij (optioneel)': 'mIw ngoq (DaH neH)',
+  'Bandgrenzen (optioneel)': '\'ay\' veHmey (DaH neH)',
   'Niet alles kon worden hersteld. Wat onleesbaar was, is bewaard gebleven.':
       'Hoch chu\'moHlaHbe\'. QIjbe\'bogh De\' polta\'.',
   'Verdieping': '\'ay\' Hurgh',